+++ Buildenator/BuilderProperties/MockingProperties.cs
@@ -1,0 +1,28 @@
+﻿using Buildenator.Abstraction;
+
+namespace Buildenator
+{
+    internal class MockingProperties
+    {
+        public MockingProperties(
+            MockingInterfacesStrategy strategy,
+            string typeDeclarationFormat,
+            string fieldDeafultValueAssigmentFormat,
+            string returnObjectFormat,
+            string[] additionalNamespaces)
+        {
+            Strategy = strategy;
+            TypeDeclarationFormat = typeDeclarationFormat;
+            FieldDeafultValueAssigmentFormat = fieldDeafultValueAssigmentFormat;
+            ReturnObjectFormat = returnObjectFormat;
+            AdditionalNamespaces = additionalNamespaces;
+        }
+
+        public MockingInterfacesStrategy Strategy { get; }
+        public string TypeDeclarationFormat { get; }
+        public string FieldDeafultValueAssigmentFormat { get; }
+        public string ReturnObjectFormat { get; }
+        public string[] AdditionalNamespaces { get; }
+
+    }
+}