+++ Tests/Buildenator.IntegrationTests.Source/AssemblyInfo.cs
@@ -1,5 +1,6 @@
-﻿using AutoFixture;
-using Buildenator.Abstraction;
+﻿using Buildenator.Abstraction;
+using Buildenator.Abstraction.Moq;
 using Buildenator.IntegrationTests.Source.Fixtures;
 
 [assembly: FixtureConfiguration(typeof(CustomFixture))]
+[assembly: MoqConfiguration()]