+++ Tests/Buildenator.IntegrationTests.Source/Builders/GranchildEntityBuilder.cs
@@ -3,6 +3,7 @@
 namespace Buildenator.IntegrationTests.Source.Builders
 {
     [MakeBuilder(typeof(GrandchildEntity))]
+    [FixtureConfiguration(typeof(AutoFixture.Fixture))]
     public partial class GrandchildEntityBuilder
     {
     }