+++ Tests/Buildenator.IntegrationTests.Source/Builders/GrandchildEntityNoMoqBuilder.cs
@@ -1,0 +1,13 @@
+﻿using Buildenator.Abstraction;
+using Buildenator.Abstraction.Moq;
+using Buildenator.IntegrationTests.Source.Fixtures;
+
+namespace Buildenator.IntegrationTests.Source.Builders
+{
+    [MakeBuilder(typeof(GrandchildEntity))]
+    [MoqConfiguration(MockingInterfacesStrategy.None)]
+    [FixtureConfiguration(typeof(CustomFixture), FixtureInterfacesStrategy.None)]
+    public partial class GrandchildEntityNoMoqBuilder
+    {
+    }
+}