+++ Tests/Buildenator.IntegrationTests.Source/IInterfaceType.cs
@@ -1,0 +1,7 @@
+﻿namespace Buildenator.IntegrationTests.Source
+{
+    public interface IInterfaceType
+    {
+        string Property { get; set; }
+    }
+}