+++ CHANGELOG.md
@@ -10,8 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## 2.0.0 - 2021-07-15
 
-## 1.2.0 - 2021-06-23
+### Added
+
+- Possibility to add a mocking fuctionality for interfaces
+  - mocking strategies introduces
+- Faking strategies
+
+
+## 1.2.0 - 2021-07-09
 
 ### Added